+++ src/model/mod.rs
@@ -1,12 +1,14 @@
 //! Model definitions for WebSocket client
 
 pub mod account;
+pub mod position;
 pub mod quote;
 pub mod subscription;
 pub mod trading;
 pub mod ws_types;
 
 pub use account::*;
+pub use position::*;
 pub use quote::*;
 pub use subscription::*;
 pub use trading::*;